systemload/requirements.txt
aiobotocore==2.3.4
aiohttp==3.7.4.post0
aioitertools==0.8.0
aioredis==1.3.1
amqp==5.1.1
appdirs==1.4.3
arabic-reshaper==2.1.3
asgiref==3.4.1
asn1crypto==1.5.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.11.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.21.21
botocore==1.24.21
cachetools==5.5.0
celery==5.2.7
certifi==2021.10.8
cffi==1.17.1
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.0
cssselect2==0.6.0
cx-Oracle==8.3.0
deprecation==2.1.0
Django==3.1.12
django-cors-headers==3.7.0
django-filter==21.1
django-nacl-fields==4.0.0
django-redis==5.2.0
django-storages==1.12.3
django-timezone-field==4.2.1
djangorestframework==3.12.4
dnspython==2.1.0
fsspec==2022.5.0
future==0.18.2
gevent==24.2.1
google-api-core==2.11.1
google-auth==2.22.0
google-cloud==0.34.0
google-cloud-vision==3.4.4
googleapis-common-protos==1.59.1
greenlet==3.1.0
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
img2pdf==0.4.4
incremental==21.3.0
jmespath==0.10.0
kombu==5.2.4
lxml==4.9.0
mock==5.1.0
multidict==6.0.5
numpy==1.21.4
opencv-python==4.10.0.84
oscrypto==1.3.0
packaging==24.1
pandas==1.3.4
payment==0.0.1.2
pdf2image==1.17.0
pikepdf==5.6.1
Pillow==8.2.0
poppler-utils==0.1.0
prompt-toolkit==3.0.23
proto-plus==1.24.0
protobuf==4.23.4
psutil==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pylightxl==1.61
PyNaCl==1.4.0
pyOpenSSL==21.0.0
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.4.2
redis==3.5.3
reportlab==3.6.9
requests==2.26.0
rsa==4.9
s3fs==2022.5.0
s3transfer==0.5.2
schedule==1.2.1
service-identity==21.1.0
six==1.15.0
speedtest-cli==2.1.3
sqlparse==0.2.4
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
txaio==21.2.1
typing-extensions==3.7.4.3
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.16.0
xhtml2pdf==0.2.7
xlrd==1.2.0
XlsxWriter==3.0.3
xlwt==1.3.0
yarl==1.11.1
zope.event==5.0
zope.interface==5.4.0